WIP to utilize SSO for VCH
Along with commit in sso-wip branch in jonathanibers/govmomi,
work on using the LoginByToken call in the Govmomi API to sign in
using a SSO token instead of user/pass.  Building on work that
@dougm did in #766.  Issues a token and uses that to log in.

I haven't figured out how to access the token granted by the Issue
command and pass it to LoginWithToken.  LWT doesn't seem to take any
parameter that could be used as a token.